refactor countdown hook
