+++ src/components/Countdown.js
@@ -1,5 +1,33 @@
-import React from 'react';
-import useCounter from '../hooks/counter-hook';
+import React, { useEffect, useState } from 'react';
+
+const calculateTimeLeft = () => {
+  let year = new Date().getFullYear();
+  const difference = +new Date(`${year}-2-15`) - +new Date();
+  let timeLeft = {};
+
+  if (difference > 0) {
+    timeLeft = {
+      days: Math.floor(difference / (1000 * 60 * 60 * 24)),
+      hours: Math.floor((difference / (1000 * 60 * 60)) % 24),
+      minutes: Math.floor((difference / 1000 / 60) % 60),
+      seconds: Math.floor((difference / 1000) % 60),
+    };
+  }
+
+  return timeLeft;
+};
+
+const useCounter = () => {
+  const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
+
+  useEffect(() => {
+    setTimeout(() => {
+      setTimeLeft(calculateTimeLeft());
+    }, 1000);
+  }, [timeLeft]);
+
+  return timeLeft;
+}
 
 export default function Countdown() {
   const timeleft = useCounter();
