--- src/hooks/counter-hook.js
@@ -1,32 +1,0 @@
-import { useEffect, useState } from "react";
-
-const calculateTimeLeft = () => {
-  let year = new Date().getFullYear();
-  const difference = +new Date(`${year}-2-15`) - +new Date();
-  let timeLeft = {};
-
-  if (difference > 0) {
-    timeLeft = {
-      days: Math.floor(difference / (1000 * 60 * 60 * 24)),
-      hours: Math.floor((difference / (1000 * 60 * 60)) % 24),
-      minutes: Math.floor((difference / 1000 / 60) % 60),
-      seconds: Math.floor((difference / 1000) % 60),
-    };
-  }
-
-  return timeLeft;
-};
-
-const useCounter = () => {
-  const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
-
-  useEffect(() => {
-    setTimeout(() => {
-      setTimeLeft(calculateTimeLeft());
-    }, 1000);
-  }, [timeLeft]);
-
-  return timeLeft;
-}
-
-export default useCounter;
